Refactor all models / routes.

+++ backend/routes/v1/services.js
@@ -5,51 +5,32 @@ const ROUTE = require('../../../shared/constants/routes');
 
 const models = require('../../models');
 
-const getFieldsFromObject = require('../../helpers/getFieldsFromObject');
+const NotFoundError = require('../../errors/not-found-error');
+
+const getFieldsFromObject = require('../../helpers/get-fields-from-object');
 
 const router = new Router({ prefix: ROUTE.BASE });
-
-const SERVICE_FIELDS = ['title', 'description'];
 
 router.get(ROUTE.SERVICES.ALL, async ctx => {
   const services = await models.Service.findAll();
 
   ctx.body = {
-    status: HTTP_STATUS.OK,
     data: services,
   };
 });
 
 router.post(ROUTE.SERVICES.ALL, async ctx => {
-  const serviceData = getFieldsFromObject(ctx.request.body, SERVICE_FIELDS);
+  const serviceData = getFieldsFromObject(ctx.request.body, models.Service.fields);
 
-  try {
-    await models.Service.create(serviceData);
-
-    ctx.body = {
-      status: HTTP_STATUS.OK,
-    };
-  } catch (error) {
-    ctx.body = {
-      status: HTTP_STATUS.WRONG_FIELDS,
-    };
-  }
+  await models.Service.create(serviceData);
 });
 
 router.delete(ROUTE.SERVICES.SINGLE, async ctx => {
-  try {
-    const service = await models.Service.findOne({ where: { id: ctx.params.id } });
-    
-    await service.destroy();
+  const service = await models.Service.findOne({ where: { id: ctx.params.id } });
 
-    ctx.body = {
-      status: HTTP_STATUS.OK,
-    };
-  } catch (error) {
-    ctx.body = {
-      status: HTTP_STATUS.NOT_FOUND,
-    };
-  }
+  if (!service) throw new NotFoundError();
+
+  await service.destroy();
 });
 
 module.exports = router;
